Fix stupid mistake with the Legendre polynomials

+++ main.cu
@@ -62,7 +62,7 @@ int main(int argc, char *argv[]) {
     double4 *d_ks;
     
     gpuErrchk(cudaMalloc((void **)&d_Bk, p.geti("num_data")*sizeof(double)));
-    gpuErrchk(cudaMalloc((void **)&d_ks, p.geti("num_data")*sizeof(float4)));
+    gpuErrchk(cudaMalloc((void **)&d_ks, p.geti("num_data")*sizeof(double4)));
     
     std::vector<double> start_params;
     std::vector<bool> limit_params;
